+++ src/phan.php
@@ -10,11 +10,6 @@ gc_disable();
 // Check the environment to make sure Phan can run successfully
 require_once(__DIR__ . '/requirements.php');
 
-// Build a code base based on PHP internally defined
-// functions, methods and classes before loading our
-// own
-$code_base = require(__DIR__ . '/codebase.php');
-
 require_once(__DIR__ . '/Phan/Bootstrap.php');
 
 use Phan\CLI;
@@ -22,6 +17,12 @@ use Phan\Phan;
 
 // Create our CLI interface and load arguments
 $cli = CLI::fromArgv();
+
+// Build a code base based after parsing the configuration,
+// so that included_extension_subset will work.
+//
+// Phan filters out user-defined functions/classes/constants.
+$code_base = require(__DIR__ . '/codebase.php');
 
 // Analyze the file list provided via the CLI
 $is_issue_found =